refactor: extract gen_benchmarks_common.py.
